Backend refactor
