+++ src/cpu/cpu-device.cpp
@@ -9,7 +9,7 @@ namespace rhi::cpu {
 
 DeviceImpl::~DeviceImpl() {}
 
-Result DeviceImpl::initialize(const DeviceDesc& desc)
+Result DeviceImpl::initialize(const DeviceDesc& desc, BackendImpl* backend)
 {
     SLANG_RETURN_ON_FAIL(Device::initialize(desc));
 
@@ -139,31 +139,3 @@ void DeviceImpl::customizeShaderObject(ShaderObject* shaderObject)
 }
 
 } // namespace rhi::cpu
-
-namespace rhi {
-
-IAdapter* getCPUAdapter(uint32_t index)
-{
-    static Adapter adapter = []()
-    {
-        Adapter outAdapter;
-        AdapterInfo info = {};
-        info.deviceType = DeviceType::CPU;
-        info.adapterType = AdapterType::Software;
-        string::copy_safe(info.name, sizeof(info.name), "Default");
-        outAdapter.m_info = info;
-        outAdapter.m_isDefault = true;
-        return outAdapter;
-    }();
-    return index == 0 ? &adapter : nullptr;
-}
-
-Result createCPUDevice(const DeviceDesc* desc, IDevice** outDevice)
-{
-    RefPtr<cpu::DeviceImpl> result = new cpu::DeviceImpl();
-    SLANG_RETURN_ON_FAIL(result->initialize(*desc));
-    returnComPtr(outDevice, result);
-    return SLANG_OK;
-}
-
-} // namespace rhi
