+++ src/d3d11/d3d11-device.cpp
@@ -1,5 +1,6 @@
 #define _CRT_SECURE_NO_WARNINGS
 #include "d3d11-device.h"
+#include "d3d11-backend.h"
 #include "d3d11-buffer.h"
 #include "d3d11-utils.h"
 #include "d3d11-query.h"
@@ -18,41 +19,11 @@
 
 namespace rhi::d3d11 {
 
-inline Result getAdaptersImpl(std::vector<AdapterImpl>& outAdapters)
-{
-    std::vector<ComPtr<IDXGIAdapter>> dxgiAdapters;
-    SLANG_RETURN_ON_FAIL(enumAdapters(dxgiAdapters));
-
-    for (const auto& dxgiAdapter : dxgiAdapters)
-    {
-        AdapterInfo info = getAdapterInfo(dxgiAdapter);
-        info.deviceType = DeviceType::D3D11;
-
-        AdapterImpl adapter;
-        adapter.m_info = info;
-        adapter.m_dxgiAdapter = dxgiAdapter;
-        outAdapters.push_back(adapter);
-    }
-
-    // Mark default adapter (prefer discrete if available).
-    markDefaultAdapter(outAdapters);
-
-    return SLANG_OK;
-}
-
-std::vector<AdapterImpl>& getAdapters()
-{
-    static std::vector<AdapterImpl> adapters;
-    static Result initResult = getAdaptersImpl(adapters);
-    SLANG_UNUSED(initResult);
-    return adapters;
-}
-
 DeviceImpl::DeviceImpl() {}
 
 DeviceImpl::~DeviceImpl() {}
 
-Result DeviceImpl::initialize(const DeviceDesc& desc)
+Result DeviceImpl::initialize(const DeviceDesc& desc, BackendImpl* backend)
 {
     SLANG_RETURN_ON_FAIL(Device::initialize(desc));
 
@@ -88,7 +59,7 @@ Result DeviceImpl::initialize(const DeviceDesc& desc)
     m_dxgiFactory = getDXGIFactory(getRHI()->getDebugLayerOptions(), this);
 
     AdapterImpl* adapter = nullptr;
-    SLANG_RETURN_ON_FAIL(selectAdapter(this, getAdapters(), desc, adapter));
+    SLANG_RETURN_ON_FAIL(selectAdapter(this, backend->getAdapters(), desc, adapter));
     m_dxgiAdapter = adapter->m_dxgiAdapter;
 
     // We will ask for the highest feature level that can be supported.
@@ -665,21 +636,3 @@ Result DeviceImpl::createRootShaderObjectLayout(
 }
 
 } // namespace rhi::d3d11
-
-namespace rhi {
-
-IAdapter* getD3D11Adapter(uint32_t index)
-{
-    std::vector<d3d11::AdapterImpl>& adapters = d3d11::getAdapters();
-    return index < adapters.size() ? &adapters[index] : nullptr;
-}
-
-Result createD3D11Device(const DeviceDesc* desc, IDevice** outDevice)
-{
-    RefPtr<d3d11::DeviceImpl> result = new d3d11::DeviceImpl();
-    SLANG_RETURN_ON_FAIL(result->initialize(*desc));
-    returnComPtr(outDevice, result);
-    return SLANG_OK;
-}
-
-} // namespace rhi
