+++ src/d3d12/d3d12-base.h
@@ -14,6 +14,7 @@
 
 namespace rhi::d3d12 {
 
+class BackendImpl;
 class AdapterImpl;
 class DeviceImpl;
 class BufferImpl;
