+++ src/wgpu/wgpu-device.h
@@ -30,7 +30,7 @@ public:
     RefPtr<CommandQueueImpl> m_queue;
 
     ~DeviceImpl();
-    virtual SLANG_NO_THROW Result SLANG_MCALL initialize(const DeviceDesc& desc) override;
+    Result initialize(const DeviceDesc& desc, BackendImpl* backend);
 
     void reportError(const char* func, WGPUStringView message);
     void reportDeviceLost(WGPUDeviceLostReason reason, WGPUStringView message);
@@ -141,10 +141,3 @@ private:
 };
 
 } // namespace rhi::wgpu
-
-namespace rhi {
-
-IAdapter* getWGPUAdapter(uint32_t index);
-Result createWGPUDevice(const DeviceDesc* desc, IDevice** outDevice);
-
-} // namespace rhi
